+++ test/nativeOptions.test.ts
@@ -28,6 +28,14 @@ describe('nativeOptions', () => {
     expect(nativeOptions.enableWatchdogTerminationTracking).toBeTruthy();
   });
 
+  test('enableCaptureFailedRequests is set when defined', async () => {
+    const nativeOptions = FilterNativeOptions(
+      {
+        enableCaptureFailedRequests: true
+      });
+    expect(nativeOptions.enableCaptureFailedRequests).toBeTruthy();
+  });
+
   test('invalid types not included', async () => {
     const nativeOptions = FilterNativeOptions(
       {
@@ -52,6 +60,7 @@ describe('nativeOptions', () => {
         enableNdkScopeSync: true,
         enableOutOfMemoryTracking: true,
         enableTracing: true,
+        enableCaptureFailedRequests: true,
         environment: 'Prod',
         ignoreErrors: ['test'],
         ignoreTransactions: ['test'],